Feature Change: When you hit the save button, all masks are stored on disk. Currently the mask is stored in the execution path of the application (clmedview is in /home/stuupke/clmedview, so the masks are stored in /home/stuupke/x_clone_x.nii)
Changes:
Files are now stored in the directory of the original image. The _clone_ is replaced with _mask_. The masks are numbered with preceding zero's.
The file path is stored in the serie struct at creation of the struct. If the file exisist, the file will be regardless overwritten.

include/lib-memory-serie.h
* Changed funcion declareation memory_serie_new()

lib/lib-memory/lib-memory-io.c
source/gui/mainwindow.c
* Changed function arguments

include/lib-memory-serie.c
* Added some code to achive the desired functionality memory_serie_new()